handle status and approved images
